+++ crates/app/src/commands/run.rs
@@ -4,7 +4,7 @@ use crate::session::CliSession;
 use clap::Args;
 use iocraft::prelude::element;
 use moon_action_context::{ActionContext, ProfileType};
-use moon_action_graph::RunRequirements;
+use moon_action_graph::{ActionGraphBuilderOptions, RunRequirements};
 use moon_affected::{DownstreamScope, UpstreamScope};
 use moon_cache::CacheMode;
 use moon_common::{is_ci, is_test_env};
@@ -52,6 +52,12 @@ pub struct RunArgs {
         help = "Bypass cache and force update any existing items"
     )]
     pub update_cache: bool,
+
+    #[arg(
+        long,
+        help = "Run the task without including sync and setup related actions in the graph"
+    )]
+    pub no_actions: bool,
 
     #[arg(
         long,
@@ -150,7 +156,17 @@ pub async fn run_target(
     };
 
     // Generate a dependency graph for all the targets that need to be ran
-    let mut action_graph_builder = session.build_action_graph(&workspace_graph).await?;
+    let mut action_graph_builder = if args.no_actions {
+        session
+            .build_action_graph_with_options(
+                &workspace_graph,
+                ActionGraphBuilderOptions::new(false),
+            )
+            .await?
+    } else {
+        session.build_action_graph(&workspace_graph).await?
+    };
+
     action_graph_builder.set_touched_files(touched_files)?;
 
     if let Some(query_input) = &args.query {